// gpa_disparity_ts/src/interfaces/CourseInfo.ts
export default interface CourseInfo {
  subject: string;
  section_status: string;
  schedule_information: string;
  section_title: string;
  type_code: string;
  building: string;
  section_info: string;
  name: string;
  enrollment_status: string;
  degree_attributes: string;
  days_of_week: string;
  section: string;
  room: string;
  yearterm: string;
  section_credit_hours: string;
  term: string;
  start_time: string;
  end_time: string;
  credit_hours: string;
  status_code: string;
  number: string;
  part_of_term: string;
  year: string;
  description: string;
  type: string;
  instructors: string;
}